requirements.txt
zappa==0.46.2
pillow==5.2.0
boto3==1.7.75
elasticsearch==6.3.1
s3transfer==0.1.13
botocore==1.10.75
